third_party/WebKit/LayoutTests/fast/peerconnection/RTCRtpSender-setParameters-crash.html

<!DOCTYPE html>
<html>
<head>
<title>RTCPeerConnection.getParameters</title>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
</head>
<body>
<script>
// TODO(orphis): Move test to main test file after https://crbug.com/webrtc/9141 is fixed
// WebRTC currently DCHECKs for a minimum maxBitrate value (currently 30000 bps).
// This can accidentally be hit when mixing units (Kbps vs B/s vs bps).
promise_test(async function() {
  let pc1 = new RTCPeerConnection();
  let pc2 = new RTCPeerConnection();

  let stream = await navigator.mediaDevices.getUserMedia({video:true});
  let videoSender = pc1.addTrack(stream.getVideoTracks()[0], stream);

  await doSignalingHandshake(pc1, pc2);
  let videoParameters = videoSender.getParameters();
  videoParameters.encodings[0].maxBitrate = 50;
  await videoSender.setParameters(videoParameters);

  videoParameters = videoSender.getParameters();
  assert_equals(videoParameters.encodings[0].maxBitrate, 50);
}, 'setParameters() set low maxBitrate value');

/**
 * Helper functions to tests.
 */
async function doSignalingHandshake(localPc, remotePc) {
  let offer = await localPc.createOffer();
  await localPc.setLocalDescription(offer);
  await remotePc.setRemoteDescription(offer);

  let answer = await remotePc.createAnswer();
  await remotePc.setLocalDescription(answer);
  await localPc.setRemoteDescription(answer);
}

</script>
</body>
</html>
